requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.0
pandas==3.0.2
scikit-learn==1.8.0
langchain==1.2.15
openai==2.31.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
psycopg2-binary==2.9.11
